replace our methods with pms
